+++ packages/eslint-plugin-dialtone/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.11.0](https://github.com/dialpad/dialtone/compare/eslint-plugin-dialtone/v1.10.0...eslint-plugin-dialtone/v1.11.0) (2025-11-25)
+
+
+### Features
+
+* **Stack:** DLT-2847 add align and justify props ([#979](https://github.com/dialpad/dialtone/issues/979)) ([539ff24](https://github.com/dialpad/dialtone/commit/539ff24c61d405a23cf6bb541a17b81d48d38feb))
+
 # [1.10.0](https://github.com/dialpad/dialtone/compare/eslint-plugin-dialtone/v1.9.0...eslint-plugin-dialtone/v1.10.0) (2025-06-17)